Assign permission to delete and modify the review post

+++ php/review/get_review_html.php
@@ -105,10 +105,21 @@
                         <div class="hit"><?=$hit?></div>
                     </div>
                     <div class="buttons">
+                <?php
+                    if ($_SESSION["usernickname"] == $user) {?>
                         <input type="button" class="button list" value="See List" onclick="location.href='list_review_html.php?page=<?=$page?>'">
                         <input type="button" class="button modify" value="Modify" onclick="location.href='modify_review_html.php?id=<?=$id?>&page=<?=$page?>'">
                         <input type="button" class="button delete" value="Delete" onclick="location.href='delete_review.php?id=<?=$id?>&page=<?=$page?>'">
                         <input type="button" class="button write" value="Write a Review" onclick="location.href='create_review_html.php'">
+                    <?php
+                    }
+                    else {
+                    ?>
+                        <input type="button" class="button list" value="See List" onclick="location.href='list_review_html.php?page=<?=$page?>'">
+                        <input type="button" class="button write" value="Write a Review" onclick="location.href='create_review_html.php'">
+                    <?php
+                    }
+                    ?>
                     </div>
                 </div>
             </div>
